-[POC] Update the application code to a single controller using hexagonal architecture.
