+++ POCCucumber/dependencies.gradle
@@ -32,7 +32,7 @@ dependencies {
     annotationProcessor 'org.projectlombok:lombok:1.18.24' // Use the latest version
     implementation 'com.google.code.findbugs:jsr305:3.0.2'
 //    implementation 'net.logstash.logback:logstash-logback-encoder:4.11' // Configurable logging
-//    implementation 'javax.validation:validation-api:2.0.1.Final' // Validation annotations
+    implementation 'javax.validation:validation-api:2.0.1.Final' // Validation annotations
 //    implementation 'commons-codec:commons-codec:1.10' // Base64 decoding
 //    implementation 'org.apache.commons:commons-lang3:3.9'
     // - R E S T
